requirements.txt
fastapi
uvicorn[standard]
pymongo
certifi
pydantic-settings
python-dotenv
requests
openai
email-validator
argon2-cffi
PyJWT
google-auth
